+++ src/matslise2d/etaIntegrals.h
@@ -2,8 +2,8 @@
 // Created by toon on 5/19/20.
 //
 
-#ifndef MATSLISE_SECTOR_CORRECTION_POTENTIAL_H
-#define MATSLISE_SECTOR_CORRECTION_POTENTIAL_H
+#ifndef MATSLISE_ETAINTEGRALS_H
+#define MATSLISE_ETAINTEGRALS_H
 
 #include "../util/calculateEta.h"
 
@@ -125,13 +125,6 @@ eta_integrals(const Scalar &delta, const Scalar &dZ1, const Scalar &dZ2) {
                 integrateTaylorTaylor(delta, dZ1, CalculateEtaData<Scalar>::taylor_eta9, dZ2,
                                       CalculateEtaData<Scalar>::taylor_eta9, CalculateEtaData<Scalar>::taylor_degree);
 
-        /*
-        std::cout << step(0, 0).transpose() << std::endl;
-        std::cout << step(0, 1).transpose() << std::endl;
-        std::cout << step(1, 0).transpose() << std::endl;
-        std::cout << step(1, 1).transpose() << std::endl;
-        */
-
         for (int i = 8; i >= 0; --i) {
             {
                 for (int j = 0; j < 2; ++j) {
@@ -156,21 +149,7 @@ eta_integrals(const Scalar &delta, const Scalar &dZ1, const Scalar &dZ2) {
                 if (i < MATSLISE_ETA_delta && j < MATSLISE_ETA_delta)
                     I(i, j) = stepUp(0);
             }
-            /*
-            I(i, j).template bottomRows<MATSLISE_INTEGRATE_delta - 2>()
-                    = ((I(i, j - 2) - (2 * j - 3) * I(i, j - 1)) / dZ2)
-                    .template topRows<MATSLISE_INTEGRATE_delta - 2>();
-             */
         }
-/*
-        std::cout << "\n****\n" << delta << ", " << dZ1 << ", " << dZ2 << std::endl;
-
-        Eigen::Array<Scalar, MATSLISE_ETA_delta * MATSLISE_ETA_delta, MATSLISE_INTEGRATE_delta> coutI;
-        for (int i = 0; i < MATSLISE_ETA_delta; ++i)
-            for (int j = 0; j < MATSLISE_ETA_delta; ++j)
-                coutI.row(i * MATSLISE_ETA_delta + j) = I(i, j);
-        std::cout << coutI << std::endl;
-*/
         return I;
     } else if (abs(Z1) < 0.5) {
         Eigen::Array<Eigen::Array<Scalar, MATSLISE_INTEGRATE_delta, 1>, 2, 2> step;
@@ -212,16 +191,6 @@ eta_integrals(const Scalar &delta, const Scalar &dZ1, const Scalar &dZ2) {
             }
         }
 
-        /*{
-            std::cout << "\n****\n" << delta << ", " << dZ1 << ", " << dZ2 << std::endl;
-
-            Eigen::Array<Scalar, MATSLISE_ETA_delta * MATSLISE_ETA_delta, MATSLISE_INTEGRATE_delta> coutI;
-            for (int i = 0; i < MATSLISE_ETA_delta; ++i)
-                for (int j = 0; j < MATSLISE_ETA_delta; ++j)
-                    coutI.row(i * MATSLISE_ETA_delta + j) = I(i, j);
-            std::cout << coutI << std::endl;
-            std::cout << std::endl;
-        }*/
         return I;
     } else if (abs(Z2) < 0.5) {
         Eigen::Array<Eigen::Array<Scalar, MATSLISE_INTEGRATE_delta, 1>, 2, 2> step;
@@ -260,33 +229,11 @@ eta_integrals(const Scalar &delta, const Scalar &dZ1, const Scalar &dZ2) {
                             .template head<MATSLISE_INTEGRATE_delta - 2>();
                     step(0, j) = curr;
                 }
-                /*
-                Eigen::Array<Scalar, MATSLISE_INTEGRATE_delta, 1> curr = stepUp(1);
-                stepUp(1)(0) = 0;
-                stepUp(1)(1) = 0;
-                stepUp(1).template tail<MATSLISE_INTEGRATE_delta - 2>()
-                        = ((stepUp(0) - (2 * j + 1) * curr) / dZ2)
-                        .template head<MATSLISE_INTEGRATE_delta - 2>();
-                stepUp(0) = curr;
-                 */
             }
         }
 
-        /*
-         * {
-            std::cout << "\n****\n" << delta << ", " << dZ1 << ", " << dZ2 << std::endl;
-
-            Eigen::Array<Scalar, MATSLISE_ETA_delta * MATSLISE_ETA_delta, MATSLISE_INTEGRATE_delta> coutI;
-            for (int i = 0; i < MATSLISE_ETA_delta; ++i)
-                for (int j = 0; j < MATSLISE_ETA_delta; ++j)
-                    coutI.row(i * MATSLISE_ETA_delta + j) = I(i, j);
-            std::cout << coutI << std::endl;
-            std::cout << std::endl;
-        }
-         */
         return I;
-    } // else
-    {
+    } else {
         Scalar I000, I101, I011, I112;
 
         if (equal) {
@@ -339,109 +286,24 @@ eta_integrals(const Scalar &delta, const Scalar &dZ1, const Scalar &dZ2) {
 
             dn /= delta;
         }
-    }
 
 
-    for (int i = 0; i < MATSLISE_ETA_delta; ++i) {
-        if (i > 1) {
-            for (int j = 0; j < 2; ++j)
+        for (int i = 0; i < MATSLISE_ETA_delta; ++i) {
+            if (i > 1) {
+                for (int j = 0; j < 2; ++j)
+                    I(i, j).template bottomRows<MATSLISE_INTEGRATE_delta - 2>()
+                            = ((I(i - 2, j) - (2 * i - 3) * I(i - 1, j)) / dZ1)
+                            .template topRows<MATSLISE_INTEGRATE_delta - 2>();
+            }
+            for (int j = 2; j < MATSLISE_ETA_delta; ++j) {
                 I(i, j).template bottomRows<MATSLISE_INTEGRATE_delta - 2>()
-                        = ((I(i - 2, j) - (2 * i - 3) * I(i - 1, j)) / dZ1)
+                        = ((I(i, j - 2) - (2 * j - 3) * I(i, j - 1)) / dZ2)
                         .template topRows<MATSLISE_INTEGRATE_delta - 2>();
-        }
-        for (int j = 2; j < MATSLISE_ETA_delta; ++j) {
-            I(i, j).template bottomRows<MATSLISE_INTEGRATE_delta - 2>()
-                    = ((I(i, j - 2) - (2 * j - 3) * I(i, j - 1)) / dZ2)
-                    .template topRows<MATSLISE_INTEGRATE_delta - 2>();
-        }
-    }
-
-
-/*
-    {
-        std::cout << "\n****\n" << delta << ", " << dZ1 << ", " << dZ2 << std::endl;
-
-        Eigen::Array<Scalar, MATSLISE_ETA_delta * MATSLISE_ETA_delta, MATSLISE_INTEGRATE_delta> coutI;
-        for (int i = 0; i < MATSLISE_ETA_delta; ++i)
-            for (int j = 0; j < MATSLISE_ETA_delta; ++j)
-                coutI.row(i * MATSLISE_ETA_delta + j) = I(i, j);
-        std::cout << coutI << std::endl;
-        std::cout << std::endl;
-    }
-    */
-
-    return I;
-}
-
-template<typename Scalar, int N = MATSLISE_N>
-void legendreTransform(const Scalar &delta, Eigen::Array<Scalar, N, 1> &quadratures) {
-    static Eigen::Matrix<Scalar, N, N> legendrePolynomials = (
-            Eigen::Matrix<Scalar, MATSLISE_N, MATSLISE_N>()
-                    << 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-                    -1, 2, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-                    1, -6, 6, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-                    -1, 12, -30, 20, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-                    1, -20, 90, -140, 70, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-                    -1, 30, -210, 560, -630, 252, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-                    1, -42, 420, -1680, 3150, -2772, 924, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-                    -1, 56, -756, 4200, -11550, 16632, -12012, 3432, 0, 0, 0, 0, 0, 0, 0, 0,
-                    1, -72, 1260, -9240, 34650, -72072, 84084, -51480, 12870, 0, 0, 0, 0, 0, 0, 0,
-                    -1, 90, -1980, 18480, -90090, 252252, -420420, 411840, -218790, 48620, 0, 0, 0, 0, 0, 0,
-                    1, -110, 2970, -34320, 210210, -756756, 1681680, -2333760, 1969110, -923780, 184756, 0, 0, 0, 0, 0,
-                    -1, 132, -4290, 60060, -450450, 2018016, -5717712, 10501920, -12471030, 9237800, -3879876, 705432, 0, 0, 0, 0,
-                    1, -156, 6006, -100100, 900900, -4900896, 17153136, -39907296, 62355150, -64664600, 42678636, -16224936, 2704156, 0, 0, 0,
-                    -1, 182, -8190, 160160, -1701700, 11027016, -46558512, 133024320, -261891630, 355655300, -327202876, 194699232, -67603900, 10400600, 0, 0,
-                    1, -210, 10920, -247520, 3063060, -23279256, 116396280, -399072960, 960269310, -1636014380, 1963217256, -1622493600, 878850700, -280816200, 40116600, 0,
-                    -1, 240, -14280, 371280, -5290740, 46558512, -271591320, 1097450640, -3155170590, 6544057520, -9816086280, 10546208400, -7909656300, 3931426800, -1163381400, 155117520
-    ).finished().template topLeftCorner<N, N>();
-
-    Scalar d = 1;
-    for (int i = 0; i < N; ++i, d /= delta)
-        quadratures(i) *= d;
-    quadratures = (legendrePolynomials * quadratures.matrix()).array();
-}
-
-template<typename Scalar, bool equal = false>
-Eigen::Array<Scalar, MATSLISE_N, 1> vbar_formulas(
-        const Eigen::Array<Scalar, MATSLISE_ETA_delta, MATSLISE_HMAX_delta> &y1,
-        const Eigen::Array<Scalar, MATSLISE_ETA_delta, MATSLISE_HMAX_delta> &y2,
-        const Scalar &delta, const Scalar &dZ1, const Scalar &dZ2) {
-
-    Eigen::Array<Eigen::Array<Scalar, MATSLISE_INTEGRATE_delta, 1>, MATSLISE_ETA_delta, MATSLISE_ETA_delta>
-            eta = eta_integrals<Scalar, equal>(delta, dZ1, dZ2);
-
-    Eigen::Array<Scalar, MATSLISE_N, 1> quadratures = Eigen::Array<Scalar, MATSLISE_N, 1>::Zero();
-/*
-    std::cout << "\ny1: " << std::endl;
-    std::cout << y1 << std::endl;
-    std::cout << "\ny2: " << std::endl;
-    std::cout << y2 << std::endl;
-*/
-
-    for (int i = 0; i < MATSLISE_ETA_delta; ++i) {
-        for (int ni = (i == 0 ? 0 : 2 * i - 1); ni < (i == 0 ? 1 : MATSLISE_HMAX_delta); ++ni) {
-            for (int j = 0; j < MATSLISE_ETA_delta; ++j) {
-                for (int nj = (j == 0 ? 0 : 2 * j - 1);
-                     (j == 0 ? nj < 1 : nj < MATSLISE_HMAX_delta) && ni + nj < MATSLISE_INTEGRATE_delta; ++nj) {
-                    Scalar s = y1(i, ni) * y2(j, nj);
-                    int count = std::min(MATSLISE_N, MATSLISE_INTEGRATE_delta - ni - nj);
-                    quadratures.segment(0, count) += s * eta(i, j).segment(ni + nj, count);
-                }
             }
         }
+
+        return I;
     }
-
-    //std::cout << quadratures.transpose() << std::endl;
-
-    legendreTransform(delta, quadratures);
-    return quadratures;
 }
 
-template<typename Scalar>
-Eigen::Array<Scalar, MATSLISE_N, 1> vbar_formulas(
-        const Eigen::Array<Scalar, MATSLISE_ETA_delta, MATSLISE_HMAX_delta> &y,
-        const Scalar &delta, const Scalar &dZ) {
-    return vbar_formulas<Scalar, true>(y, y, delta, dZ, dZ);
-}
-
-#endif //MATSLISE_SECTOR_CORRECTION_POTENTIAL_H
+#endif //MATSLISE_ETAINTEGRALS_H
