+++ src/pyslise/pyslise2d.cpp
@@ -259,14 +259,6 @@ The next set of parameters are more advanced and can be useful to tweak when the
                     l->at(i) = s.eigenvalues[i];
                 return l;
             })
-            .def_property_readonly("eigenfunctions", [](Matslise2D<>::Sector &s) -> vector<ArrayXd> * {
-                auto l = new vector<ArrayXd>(s.se2d->N);
-                for (unsigned long i = 0; i < l->size(); ++i)
-                    l->at(i) = s.eigenfunctions[i];
-                return l;
-            })
-            .def_readonly("vbar", &Matslise2D<>::Sector::vbar)
-            .def("calculateDeltaV", &Matslise2D<>::Sector::calculateDeltaV)
             .def_readonly("matslise", &Matslise2D<>::Sector::matslise, py::return_value_policy::reference)
             .def_readonly("matscs", &Matslise2D<>::Sector::matscs, py::return_value_policy::reference)
             .def_readonly("min", &Matslise2D<>::Sector::min)
